Birthday Invite w/ HTML Only
The first project I created using HTML only. It is an birthday invitation for a small party on my birthday. Obviously, I added a few extra things to make it a bit different from the original project I was instructed to create to give my touch, but the Image Source and href Link remain the same. I've also added an additional screenshot of what the final project originally looks like, so it doesn't have to be run to view what the project is.